Solves Flex 04
I opted to go for a slighlty different solution, using additional flex
elements and aligning my content without using the justify-content:center
property. I did this mostly because the brief was unclear on if the design should
be responsive to fill the whole page on a wider screen.

The end result is my implementation is more respensive but the content is more spread out
on a wider display.

Signed-off-by: Kykneon <158997169+Kykneon@users.noreply.github.com>