Solves Flex 07-02
I struggled with setting the width & height dimensions of the card
elements at first.

Specifically the vertical dimensions of the cards were behaving in
an unexpected way becuase I had their parent flex box align-items
set to center, which not only centered the items vertically, as expected
but also shrunk thier vertical dimensions to the size of their content.